+++ packages/path_provider/path_provider_linux/test/path_provider_linux_test.dart
@@ -7,7 +7,7 @@ import 'package:path_provider_platform_interface/path_provider_platform_interfac
 
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
-  PathProviderLinux.register();
+  PathProviderLinux.registerWith();
 
   setUp(() {});