+++ packages/shared_preferences/shared_preferences_linux/example/integration_test/shared_preferences_test.dart
@@ -9,6 +9,7 @@ import 'dart:async';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
 import 'package:shared_preferences_linux/shared_preferences_linux.dart';
+import 'package:shared_preferences_platform_interface/shared_preferences_platform_interface.dart';
 
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
@@ -30,14 +31,18 @@ void main() {
       'flutter.List': <String>['baz', 'quox'],
     };
 
-    SharedPreferencesLinux preferences;
+    SharedPreferencesStorePlatform preferences;
 
-    setUp(() async {
-      preferences = SharedPreferencesLinux.instance;
+    setUp(() {
+      preferences = SharedPreferencesStorePlatform.instance;
     });
 
-    tearDown(() {
-      preferences.clear();
+    tearDown(() async {
+      await preferences.clear();
+    });
+
+    test('instance is Linux implementation', () {
+      expect(preferences, isA<SharedPreferencesLinux>());
     });
 
     testWidgets('reading', (WidgetTester _) async {