+++ packages/shared_preferences/shared_preferences_linux/lib/shared_preferences_linux.dart
@@ -16,8 +16,10 @@ import 'package:shared_preferences_platform_interface/shared_preferences_platfor
 ///
 /// This class implements the `package:shared_preferences` functionality for Linux.
 class SharedPreferencesLinux extends SharedPreferencesStorePlatform {
-  /// The default instance of [SharedPreferencesLinux] to use.
-  static SharedPreferencesLinux instance = SharedPreferencesLinux();
+  /// Registers the Linux implementation.
+  static void registerWith() {
+    SharedPreferencesStorePlatform.instance = SharedPreferencesLinux();
+  }
 
   /// Local copy of preferences
   Map<String, Object>? _cachedPreferences;