+++ packages/shared_preferences/shared_preferences_macos/example/integration_test/shared_preferences_test.dart
@@ -31,7 +31,7 @@ void main() {
 
     SharedPreferencesStorePlatform preferences;
 
-    setUp(() async {
+    setUp(() {
       preferences = SharedPreferencesStorePlatform.instance;
     });