+++ packages/shared_preferences/shared_preferences_windows/example/lib/main.dart
@@ -7,7 +7,7 @@
 import 'dart:async';
 
 import 'package:flutter/material.dart';
-import 'package:shared_preferences_windows/shared_preferences_windows.dart';
+import 'package:shared_preferences_platform_interface/shared_preferences_platform_interface.dart';
 
 void main() {
   runApp(MyApp());
@@ -31,7 +31,7 @@ class SharedPreferencesDemo extends StatefulWidget {
 }
 
 class SharedPreferencesDemoState extends State<SharedPreferencesDemo> {
-  final prefs = SharedPreferencesWindows.instance;
+  final prefs = SharedPreferencesStorePlatform.instance;
   late Future<int> _counter;
 
   Future<void> _incrementCounter() async {