+++ packages/shared_preferences/shared_preferences_linux/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.0.1
+
+* Add `implements` to the pubspec.
+* Add `registerWith` to the Dart main class.
+
 ## 2.0.0
 
 * Migrate to null-safety.